cgroups: Add stubs for non-Linux platforms
